IOS-6893 Update ValidatorView
